-- Ajustados os nomes das tabelas nas Entidades;

+++ src/main/resources/db.migrations/V1__create-table-tb_users.sql
@@ -4,10 +4,10 @@ CREATE TABLE tb_users (
     email VARCHAR(255) NOT NULL,
     password VARCHAR(150) NOT NULL,
     role VARCHAR(150) DEFAULT 'USER' NOT NULL,
-    account_non_expired TINYINT(1) DEFAULT 1 NOT NULL,
-    credentials_non_expired TINYINT(1) DEFAULT 1 NOT NULL,
-    account_non_locked TINYINT(1) DEFAULT 1 NOT NULL,
-    enabled TINYINT(1) DEFAULT 1 NOT NULL,
+    account_non_expired INT DEFAULT 1 NOT NULL,
+    credentials_non_expired INT DEFAULT 1 NOT NULL,
+    account_non_locked INT DEFAULT 1 NOT NULL,
+    enabled INT DEFAULT 1 NOT NULL,
     PRIMARY KEY (user_id),
     UNIQUE (email)
 );
